closes #53
I believe this is the right syntax needed for the js equivalent of
String.equals(strOther).

Line 2076 in ProposalsManage.js reads
if (_proposalId === "0") {

The program now executes w/o the firebug error from the .equals
implementation.  
The usage of == in other equality statements may cause issues and should
be evaluated.